requirements.txt
spotipy
google-auth
google-auth-oauthlib
google-auth-httplib2
googleapiclient
pandas
python-dotenv